readd the virtual attribute
the virtual attribute is required to get the additional information
from the database. The reason why it failed when getting an object was
because lazy loading doesn't work on through WCF services.

To eagerly load the properties of an instance we have to use .include
in the context. In addition the ProxyCreationEnabled = false is
required because it will fail again. I'm not sure yet why exactly.